moved ops to operations
