+++ src/rook/operations/average.py
