+++ src/rook/operations/consolidate.py
